+++ backend/mo/ic_vetkeys/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- BREAKING CHANGE: Fixed a few inconsistencies with the Rust backend of encrypted maps. 
+
 ### Adds
 - Sign with BLS and VetKD helper functions.